Move static method to Utilities class
