+++ src/main/java/tracing/SNT.java
@@ -1,5 +1,10 @@
 package tracing;
 
+import java.awt.Font;
+import java.awt.Insets;
+
+import javax.swing.JButton;
+
 import org.scijava.util.VersionUtils;
 
 import ij.IJ;
@@ -21,6 +26,17 @@ public class SNT {
 
 	protected static void log(final String string) {
 		System.out.println("[SNT] " + string);
+	}
+
+	protected static JButton smallButton(final String text) {
+		final double SCALE = .85;
+		final JButton button = new JButton(text);
+		final Font font = button.getFont();
+		button.setFont(font.deriveFont((float) (font.getSize() * SCALE)));
+		final Insets insets = button.getMargin();
+		button.setMargin(new Insets((int) (insets.top * SCALE), (int) (insets.left * SCALE),
+				(int) (insets.bottom * SCALE), (int) (insets.right * SCALE)));
+		return button;
 	}
 
 }
